feat: add quiz router
